tbh idek what i fixed

+++ src/Controller/Actions/Editing/UnindentAction.cpp
@@ -10,12 +10,14 @@ void UnindentAction::apply(ExecutionContext& context) {
     EditorState& state = context.state;
 
     int leading_spaces = state.getParagraph(m_row).find_first_not_of(' ');
-    m_spaces_removed = leading_spaces % m_max_indent_width;
+    m_spaces_removed = std::min(leading_spaces, m_max_indent_width);
 
-    state.deleteRange(
-        {m_row, 0},
-        {m_row, m_spaces_removed - 1}
-    );
+    if (m_spaces_removed > 0) {
+        state.deleteRange(
+            {m_row, 0},
+            {m_row, m_spaces_removed - 1}
+        );
+    } 
     
     for (int _ = 0; _ < m_spaces_removed; _++) {
         state.moveCursorLeft();
